refactor try
