+++ proxy_sanity_test.go
@@ -83,70 +83,6 @@ func runInsertFindUpdate(host string, proxyPort, iteration int, shouldFail bool,
 	return nil
 }
 
-func enableFailPointCloseConnection(host string, mongoPort int, mode MongoConnectionMode) error {
-	client, err := getTestClient(host, mongoPort, mode, false, "enableFailPointCloseConnection")
-	if err != nil {
-		return err
-	}
-	ctx, cancelFunc := context.WithTimeout(context.Background(), ClientTimeoutSec)
-	defer cancelFunc()
-	if err := client.Connect(ctx); err != nil {
-		return fmt.Errorf("cannot connect to server. err: %v", err)
-	}
-	defer client.Disconnect(ctx)
-	// cannot fail "find" because it'll prevent certain driver machinery when operating against replica sets to work properly
-	cmd := bson.D{
-		{"configureFailPoint", "failCommand"},
-		{"mode", "alwaysOn"},
-		{"data", bson.D{
-			{"failCommands", []string{"update"}},
-			{"closeConnection", true},
-		}},
-	}
-	return client.Database("admin").RunCommand(ctx, cmd).Err()
-}
-
-func enableFailPointErrorCode(host string, mongoPort, errorCode int, mode MongoConnectionMode) error {
-	client, err := getTestClient(host, mongoPort, mode, false, "enableFailPointErrorCode")
-	if err != nil {
-		return err
-	}
-	ctx, cancelFunc := context.WithTimeout(context.Background(), ClientTimeoutSec)
-	defer cancelFunc()
-	if err := client.Connect(ctx); err != nil {
-		return fmt.Errorf("cannot connect to server. err: %v", err)
-	}
-	defer client.Disconnect(ctx)
-	// cannot fail "find" because it'll prevent certain driver machinery when operating against replica sets to work properly
-	cmd := bson.D{
-		{"configureFailPoint", "failCommand"},
-		{"mode", "alwaysOn"},
-		{"data", bson.D{
-			{"failCommands", []string{"update"}},
-			{"errorCode", errorCode},
-		}},
-	}
-	return client.Database("admin").RunCommand(ctx, cmd).Err()
-}
-
-func disableFailPoint(host string, mongoPort int, mode MongoConnectionMode) error {
-	client, err := getTestClient(host, mongoPort, mode, false, "disableFailPoint")
-	if err != nil {
-		return err
-	}
-	ctx, cancelFunc := context.WithTimeout(context.Background(), ClientTimeoutSec)
-	defer cancelFunc()
-	if err := client.Connect(ctx); err != nil {
-		return fmt.Errorf("cannot connect to server. err: %v", err)
-	}
-	defer client.Disconnect(ctx)
-	cmd := bson.D{
-		{"configureFailPoint", "failCommand"},
-		{"mode", "off"},
-	}
-	return client.Database("admin").RunCommand(ctx, cmd).Err()
-}
-
 func runSanityOps(host string, proxyPort, parallelism int, shouldFail bool, t *testing.T, mode MongoConnectionMode, secondaryReads bool) int32 {
 	var wg sync.WaitGroup
 	var failing int32
